requirements.txt
streamlit>=1.33
streamlit-quill>=0.1.0
pandas>=1.5
gspread>=5.8.0
google-auth>=2.27.0
google-generativeai>=0.6.0
beautifulsoup4>=4.12
markdown>=3.5
python-docx>=1.0.1

# Optional similarity backends (install both if you can)
sentence-transformers>=2.6.1
scikit-learn>=1.4
